Added partials support.
CodeKit partials can be flagged by including a trailing underscore (_) in the file name. If these files are compiled directly, the build may fail if, for example, there are variables referenced in the partial that are defined in the parent Kit file. This update checks for partials and makes sure they are not compiled directly. (See the “The ‘Partial’ Convention” section at http://incident57.com/codekit/help.html#kit)